+++ apps/parsley/src/hooks/useSections/testData.ts
@@ -1,0 +1,91 @@
+import { SectionData } from "./utils";
+
+const step = "1 of 4";
+const sectionData: SectionData = {
+  commands: [
+    {
+      commandID: "command-1",
+      commandName: "c1",
+      functionID: "function-1",
+      range: {
+        end: 6,
+        start: 1,
+      },
+      step,
+    },
+    {
+      commandID: "command-6",
+      commandName: "c2",
+      functionID: "function-1",
+      range: {
+        end: 8,
+        start: 6,
+      },
+      step,
+    },
+    {
+      commandID: "command-9",
+      commandName: "c3",
+      functionID: "function-9",
+      range: {
+        end: 12,
+        start: 9,
+      },
+      step,
+    },
+    {
+      commandID: "command-12",
+      commandName: "c4",
+      functionID: "function-9",
+      range: {
+        end: 14,
+        start: 12,
+      },
+      step,
+    },
+  ],
+  functions: [
+    {
+      functionID: "function-1",
+      functionName: "f-1",
+      range: {
+        end: 8,
+        start: 1,
+      },
+    },
+    {
+      functionID: "function-9",
+      functionName: "f-2",
+      range: {
+        end: 14,
+        start: 9,
+      },
+    },
+  ],
+};
+const initialSectionState = {
+  "function-1": {
+    commands: {
+      "command-1": {
+        isOpen: false,
+      },
+      "command-6": {
+        isOpen: false,
+      },
+    },
+    isOpen: false,
+  },
+  "function-9": {
+    commands: {
+      "command-9": {
+        isOpen: false,
+      },
+      "command-12": {
+        isOpen: false,
+      },
+    },
+    isOpen: false,
+  },
+};
+
+export { sectionData, initialSectionState };
