+++ apps/parsley/src/hooks/useSections/useSections.test.tsx
@@ -8,6 +8,7 @@ import {
 } from "test_data/parsleySettings";
 import * as ErrorReporting from "utils/errorReporting";
 import { useSections } from ".";
+import { initialSectionState, sectionData } from "./testData";
 import * as sectionUtils from "./utils";
 
 const wrapper = ({ children }: { children: React.ReactNode }) => (
@@ -307,93 +308,7 @@ describe("useSections", () => {
     "normal log line",
     "normal log line",
   ];
-  const step = "1 of 4";
-  const sectionData: sectionUtils.SectionData = {
-    commands: [
-      {
-        commandID: "command-1",
-        commandName: "c1",
-        functionID: "function-1",
-        range: {
-          end: 6,
-          start: 1,
-        },
-        step,
-      },
-      {
-        commandID: "command-6",
-        commandName: "c2",
-        functionID: "function-1",
-        range: {
-          end: 8,
-          start: 6,
-        },
-        step,
-      },
-      {
-        commandID: "command-9",
-        commandName: "c3",
-        functionID: "function-9",
-        range: {
-          end: 12,
-          start: 9,
-        },
-        step,
-      },
-      {
-        commandID: "command-12",
-        commandName: "c4",
-        functionID: "function-9",
-        range: {
-          end: 14,
-          start: 12,
-        },
-        step,
-      },
-    ],
-    functions: [
-      {
-        functionID: "function-1",
-        functionName: "f-1",
-        range: {
-          end: 8,
-          start: 1,
-        },
-      },
-      {
-        functionID: "function-9",
-        functionName: "f-2",
-        range: {
-          end: 14,
-          start: 9,
-        },
-      },
-    ],
-  };
-  const initialSectionState = {
-    "function-1": {
-      commands: {
-        "command-1": {
-          isOpen: false,
-        },
-        "command-6": {
-          isOpen: false,
-        },
-      },
-      isOpen: false,
-    },
-    "function-9": {
-      commands: {
-        "command-9": {
-          isOpen: false,
-        },
-        "command-12": {
-          isOpen: false,
-        },
-      },
-      isOpen: false,
-    },
-  };
+
   const metadata = {
     logType: LogTypes.EVERGREEN_TASK_LOGS,
     onInitOpenSectionContainingLine: undefined,
